Adding latest additions to diag_manager methods to match PR 10 in cFMS
